+++ CHANGELOG.md
@@ -5,6 +5,7 @@ HEAD
 ----
 
 * Made `JsonBuffer` non-copyable (PR #524 by @luisrayas3)
+* Added `StaticJsonBuffer::clear()`
 
 v5.10.1
 -------